Separate types export and add node backend initialization
