+++ packages/unmock-core/src/types.ts
@@ -1,0 +1,11 @@
+export { RequestResponseSpy } from "./service/spy/index";
+export {
+  ISerializedRequest as Request,
+  ISerializedRequest as UnmockRequest,
+  ISerializedResponse as UnmockResponse,
+  ServiceStoreType as UnmockServices,
+} from "./interfaces";
+export {
+  IService as UnmockService,
+  IService as Service,
+} from "./service/interfaces";
